+++ smoderp2d/providers/base/__init__.py
@@ -274,12 +274,12 @@ class BaseProvider(object):
         nodata = tf.equal(data['mat_efect_cont'], -3.40282347e+38)
         nodata_pos = tf.constant(
             [[3.40282347e+38, ] * nodata.shape[1], ] * nodata.shape[0],
-            dtype=tf.float32)
+            dtype=tf.float64)
         data['mat_efect_cont_tf'] = tf.where(nodata, nodata_pos,
                                              data['mat_efect_cont'])
         data['rr_tf'] = tf.constant(data['rr'], dtype=tf.int32)
         data['rc_tf'] = tf.constant(data['rc'], dtype=tf.int32)
-        data['mat_n_tf'] = tf.constant(data['mat_n'], dtype=tf.float32)
-        data['mat_slope_tf'] = tf.constant(data['mat_slope'], dtype=tf.float32)
+        data['mat_n_tf'] = tf.constant(data['mat_n'], dtype=tf.float64)
+        data['mat_slope_tf'] = tf.constant(data['mat_slope'], dtype=tf.float64)
 
         return data
